// src/app/core/services/model/userOutputDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuthorityTemplateNames } from './authorityTemplateNames';

export interface UserOutputDto { 
    uuid?: string;
    userId?: string;
    loginName?: string;
    lastName?: string;
    firstName?: string;
    email?: string;
    authorityTemplateName?: AuthorityTemplateNames;
    lastLoginTimestamp?: Date;
    isLocked?: boolean;
}