// src/app/core/services/model/userSetAuthorityTemplateInput.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuthorityTemplateNames } from './authorityTemplateNames';

export interface UserSetAuthorityTemplateInput { 
    managerPassword: string;
    authorityTemplateName: AuthorityTemplateNames;
}